Move client to separate file
